move sharable page part to tmp-shared
