+++ webui/react/src/tmp-shared/components/Page/index.ts
@@ -1,0 +1,3 @@
+export * from './Page';
+import Page from './Page';
+export default Page;
